--- experiments/TWPA/experiments.py
@@ -1,133 +1,0 @@
-"""Collection of experiments for characterization of a TWPA."""
-from time import sleep
-
-import numpy as np
-from twpa_experiment import TWPAExperiment
-
-from qtics import VNAN9916A
-from qtics.experiment import Experiment
-
-
-class VNASnapshot(Experiment):
-    """Simple snapshot with vna."""
-
-    vna = VNAN9916A("vna", "192.168.40.10")
-
-    def main(self):
-        """Acquire VNA snapshot."""
-        self.vna.connect()
-        self.vna.set_defaults(
-            f_min=1e9, f_max=1e10, sweep_points=10000, average=1, IFBW=1000, power=-10
-        )
-        f, z = self.vna.snapshot()
-        self.append_data_group("snapshot", datasets={"frequencies": f, "values": z})
-        self.save_config()
-
-
-class S21VsBias(TWPAExperiment):
-    """Response as a function of bias voltage."""
-
-    delay_between_acquisitions = 5
-    min_bias = 0
-    max_bias = 2.4
-    step_bias = 0.02
-
-    def main(self):
-        """Acquire for different bias voltage values."""
-        self.vna.set_defaults(power=-20)
-        self.pump.rf_status = "OFF"
-        bias_sweep = np.arange(self.min_bias, self.max_bias, self.step_bias)
-        self.attribute_sweep("bias", "voltage", bias_sweep)
-
-
-class GainVsFreq(TWPAExperiment):
-    """Gain profile."""
-
-    def main(self):
-        """Acquire with and without pump."""
-        self.pump.rf_status = "ON"
-        self.bias.output_on()
-        sleep(0.1)
-        f, z = self.vna.snapshot()
-        self.append_data_group("pump_on", datasets={"frequencies": f, "values": z})
-        self.pump.rf_status = "OFF"
-        self.bias.output_off()
-        sleep(self.delay_between_acquisitions)
-        self.bias.output_on()
-        f, z = self.vna.snapshot()
-        self.append_data_group("pump_off", datasets={"frequencies": f, "values": z})
-        self.bias.output_off()
-
-
-class GainVsPumpFreq(TWPAExperiment):
-    """Gain as a function of pump frequency."""
-
-    min_f_pump = 7.8e9
-    max_f_pump = 8.3e9
-    f_pump_step = 1e7
-
-    min_p_pump = -4
-    max_p_pump = -2
-    p_pump_step = 0.1
-
-    def main(self):
-        """Acquire for different pump frequency values."""
-        self.bias.output_on()
-        self.pump.rf_status = "OFF"
-        f, z = self.vna.snapshot()
-        self.append_data_group("pump_off", datasets={"frequencies": f, "values": z})
-        self.bias.output_off()
-
-        self.pump.rf_status = "ON"
-        freq_sweep = np.arange(self.min_f_pump, self.max_f_pump, self.f_pump_step)
-        self.attribute_sweep("pump", "frequency", freq_sweep)
-        self.pump.rf_status = "OFF"
-        self.bias.output_off()
-
-
-class GainVsPumpFreqVsPumpPow(TWPAExperiment):
-    """Gain as a function of pump frequency and power."""
-
-    min_f_pump = 7.8e9
-    max_f_pump = 8.3e9
-    f_pump_step = 1e7
-
-    min_p_pump = -4
-    max_p_pump = -2
-    p_pump_step = 0.1
-
-    def main(self):
-        """Acquire for different pump frequency values."""
-        self.bias.output_on()
-        self.pump.rf_status = "OFF"
-        f, z = self.vna.snapshot()
-        self.append_data_group("pump_off", datasets={"frequencies": f, "values": z})
-        self.bias.output_off()
-
-        for pump_pow in np.arange(self.min_p_pump, self.max_p_pump, self.p_pump_step):
-            self.pump.rf_status = "ON"
-            group_name = f"pow_{pump_pow}"
-            self.append_data_group(group_name, power=pump_pow)
-            freq_sweep = np.arange(self.min_f_pump, self.max_f_pump, self.f_pump_step)
-            self.attribute_sweep(
-                "pump", "frequency", freq_sweep, parent_name=group_name
-            )
-            self.pump.rf_status = "OFF"
-        self.bias.output_off()
-
-
-class CompressionPoint(TWPAExperiment):
-    """Gain as a function of signal power."""
-
-    f_stop = 7e9
-    min_vna_pow = -35
-    max_vna_pow = -5
-    step_vna_pow = 0.5
-
-    def main(self):
-        """Acquire for different vna power values."""
-        self.pump.rf_status = "ON"
-        power_sweep = np.arange(self.min_vna_pow, self.max_vna_pow, self.step_vna_pow)
-        self.attribute_sweep("vna", "power", power_sweep, parent_name="pump_on")
-        self.pump.rf_status = "OFF"
-        self.attribute_sweep("vna", "power", power_sweep, parent_name="pump_off")
